Refactor API to include partnerNr in company and person
